+++ backend/src/main/java/com/advocacychat/backend/dto/ChatDTO.java
@@ -6,9 +6,9 @@ import java.time.LocalDateTime;
 
 @Data
 public class ChatDTO {
-    private Integer id;
-    private Integer clienteId;
-    private Integer advogadoId;
+    private Long id;
+    private Long clienteId;
+    private Long advogadoId;
     private Boolean ativo;
     private LocalDateTime criadoEm;
 }
