+++ backend/src/main/java/com/advocacychat/backend/repository/UsuarioRepository.java
@@ -4,8 +4,9 @@ import com.advocacychat.backend.model.UsuarioModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 
+import java.util.Locale;
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<UsuarioModel, Integer> {
+public interface UsuarioRepository extends JpaRepository<UsuarioModel, Long> {
     Optional<UserDetails> findByEmail(String email);
 }
